Added InitFromFile support to Assembler module.
Also added a temporary hack to allow multiple source files to be provided on
the command line.  In the future I will pull this type of functionality out
into a separate module like Builder but for now I want to share a single
symbol table across all source files (something I don't want to do for the
final assembler) so it is best placed in the Assembler module itself.
I also cleaned up some of the test cases to remove some redundant code.