Also shunt free() call through function pointer.
CppUTest uses a macro to pull free() calls into one of its own functions which
records the file and line making the allocation.  This means that the best I
could do in production code was to override the function from CppUTest and just
call the CRT free() implementation but this still left the file and line number
being passed around.  Using a hook for free() just like malloc() allows the
production code to not even require the file and line information.