Ignore LST directive and produce error for unrecognized operators.
For now I am just ignoring the LST directive since I always want to see the
listing produced at this time because it is what I use to manually verify the
progress in addition to the automated unit tests.

I also added code which will flag an error every time an unrecognized operator
is encountered on a line.  This allows me to track progress towards getting
BOOT.S to assemble with this program.  There are currently 138 lines with
unknown operators.