Modified Symbol format and SymbolTable_Add() method.
I have removed the pData field from the Symbol data structure and the calls to
SymbolTable_Add().  This field has been changed from a void* to an unsigned
short which holds the symbol value.  During the add call, only the key value is
passed in.  The value fields are filled in by the caller into the returned
symbol pointer.