Refactored AssemblerTest module.
Implemented the test6502Instruction() method which can be passed a string
that represents a row from the 65C02 instruction table.  For supported
addressing modes, the expected opcode will be listed.  If the instruction
doesn't support an addressing mode, a XX is used instead.  Each table
entry is separated from the previous with a comma.  Sometimes a zero page
variant of an addressing mode isn't supported for an instruction but in that
case the assembler will use the absolute variant instead if it is supported.
This can be indicated in the string by placing the absolute opcode in the
zeropage entry with a prefix of '^' to indicate that a 3 byte instruction will
be generated instead of a 2 byte variant.  It should also be noted that the
accumulator and implicit entries are collapsed into a single column.

An example string for the CMP instruction:
    "C9,CD,C5,XX,C1,D1,D5,^D9,DD,D9,XX,XX,XX,D2"