Added support for BEQ, CLC, JMP, and LDX.
Test all addressing modes for these instructions.

The biggest change for this commit is the addition of code to the zero page and
absolute addressing mode handling code to check and see if relative addressing
mode should be used instead to represent the offset of the absolute address
from the current program counter value.