Modified Assembler module to get 100% code coverage.
There is a try/catch block which catches invalid addressing mode errors so that
means that ADDRESSING_MODE_INVALID never makes it to the switch below.  This
caused less than 100% code coverage for this switch case.  Removing the case
doesn't work since the compiler then notices that not all enumeration values
are covered.  I therefore just brough this case up and combined it with
the absolute addressing mode handler code.