Implemented basic DS directive support.
Doesn't currently support the '\' character as an operand to indicate the rest
of the current page should be filled or a ',' following the repeat expression
to indicate the fill value to be used.  There are tests in place to make sure
that any POP code which happens to use these would generate errors so that I
don't miss implementing if/when needed.