Refactor code to use BinaryBuffer module for LineInfo::machineCode allocations.
The previous code just had 32 bytes per LineInfo structure that could be used
to hold the resulting machine code for each line of assembly.  This was enough
to support the HEX directive but was too restrictive for other directives such
as ASC.  Using BinaryBuffer allocations also makes it easier to save the final
object file to disk as it just needs to dump all of the allocations
sequentially.

Other modifications made as part of this commit:
* DUM directives now just switch the assembler to use a separate BinaryBuffer
  reserved for such dummy sections.  The DEND directive will switch the
  assembler back to the main object code buffer.
* I also removed the special DS directive handling code as the code can now
  just allocate the number of bytes needed for the DS call and fill
  it with the necessary data.