Forward references were corrupting object buffer.
The pMachineCode array of bytes in the LineInfo structure for a line which
contained forward references was getting reallocated during the back patch
when the symbol was finally encountered.  This caused their to be gaps in the
object buffer which contained the original orphaned machine code and then the
correct machine code would be injected into the object file closer to where the
forward referenced symbol was actually defined later.

Modified the code to not allow the pMachineCode field in the LineInfo structure
to be allocated more than once.  Instead the code just checks to make sure that
the subsequent allocation attempts use the same machine code size and if not,
an error will be generated since the assembler wasn't able to infer the size
of the forward reference correctly.