Initial commit of ByteBuffer module.
This module couples a byte array and its size together in a single entity.  It
has an allocation method which will free any previous allocation so that the
buffer module can be reused easily.  It also has methods which can read/write
from/to files and will throw fileExceptions if the transfer is truncated to be
less than the size of the buffer allocation.