Added support for all remaining 65c02 instructions.
This commit also created the test6502RelativeBranchInstruction helper routine
that is able to verify that a branch instruction works as expected and reduces
the amount of redundant code in the main branch tests which verified that the
right machine code was created for the supported relative addressing mode and
the proper error was generated for all of the other addressing modes.

NOTE: The assembler doesn't currently know to return errors when using
enhanced 65c02 instructions when not in that mode.  I will add this
functionality once I implement support for the XC directive that enables this
mode.