Allow forward references to local labels occur on line which defines global label.
Previously the following code didn't work as expected:
WRITE
 ...
RWSECTS bne :noncont
 ...
:noncont
 ...

The assembler thought that the :noncont local label should be found within the
scope of the previous WRITE global rather than the RWSECTS which was defines on
the same line as the forward reference.

Previously snap would create symbols for labels when it encountered an EQU
directive or once the line was done being parsed if it wasn't already added
because of an EQU.  The order is now reversed and the symbol will be created
for a line even before parsing of the operator begins.  If it happens to be an
EQU directive then the symbol will be updated to use the EQU expression
operand.  This required refactoring some code to still correctly fixup forward
references and produce expected list file output for EQU lines.