Added fake support for XC directive.
Code which uses the XC directive will now assemble but it doesn't really do
anything.  One XC directive switches a mode in the assembler module so that it
knows to support 65c02 instructions but the instruction set isn't yet
restricted to 6502 instructions when a XCE directive hasn't been seen.  Two XC
directives should allow 65816 instructions but the assembler doesn't currently
support this extended instruction set so it just forces all instructions to
resolve to a rts instruction.

Note: The 6502 restriction will be fixed in a soon to come commit but the
      65816 support will be awhile in coming.  This does mean that Prince of
      Persian 65816 specific code won't work properly until this support is
      added.