Moved instruction set tables into Assembler module.
The Assembler module now has an instructionSets array of pointers to instruction
set tables for each of the 3 sets in the InstructionSetSupported enumeration.
The firstPassAssembleLine() method now uses the entry from this array which
matches the currently selected instruction set.  Currently there is only one
instruction set table and it is duplicated into all 3 of the elements of this
instructionSets array.  The source instruction set tables will be found in the
new src/InstructionSets.h header.  These tables are runtime sorted by the
pOperator string so that firstPassAssembleLine() can perform a binary search
instead of a slower linear search.

An upcoming commit will make it so that the 6502 and 65c02 tables aren't the
same but just contain the instructions supported when it is the selected set.