Improve code coverage in AddressingMode module after last change.
As far as I can tell, it was really getting 100% code coverage after my
previous commit but calling a function which only contains a code path which
ends in a __throw macro flags the call itself as not being executed even
though the bulk of the called function itself was executed according to the
same code coverage data.

I switched to using a macro for reportAndThrowOnInvalidAddressingMode() and
reportAndThrowOnInvalidIndexRegister() functionality.