ExpressionEval module only takes SizedStrings as input.
Previously there was both a ExpressionEval and ExpressionEvalSizedString
methods available to parse an expression string.  Now there is just the
ExpressionEval method but it only takes a SizedString input parameter and there
is no longer a version which takes a normal NULL terminated string.  This is
the first commit in several that I am working on to push the use of
SizedStrings throughout more of the code base and reduce number of source line
copies that need to be made.