Added SizedString_strdup() method.
Also moved SizedString module sources and tests to libcommon from libsnap.
Added this method so that a SizedString can be easily converted into NULL
terminated string.  The TextFile module was converted to use SizedStrings for
the filename since they could be sourced from PUT directives in the code but
the fopen() API requires a NULL terminated string.