Now track line number and filename in TextFile module.
This information was moved into this module from the Assembler module since
the PUT directive to be soon added will make it possible for more than one file
to be in use during the assembly process.

Added methods include:
TextFile_GetLineNumber()
TextFile_GetFilename()

The TextFile_CreateFromFile() method was also updated to take a SizedString for
the filename so that it could be more easily sourced from the PUT directive.