Modified TextFile module to use NULL terminated strings again.
I didn't realize that the PUT directive just specified the base filename and
not the .S suffix as well so it needs to be appended to the specified filename.
This makes the SizedString parsed from the source file not directly useful
anyway.  The concatenation of the base filename and suffix can now be done
in the TextFile_CreateFromFile() module to create the full filename and store
in a private member field.