Implement PUT directive.
This adds basic support for the PUT directive to the snap assembler but I
still want to add support for an INCLUDE path so that it will work better
with how I plan to build PoP sources.

Changes include:
LineInfo structure now keeps a pointer to the TextFile object from which its
source originated.  This makes error messages related to the line point to
the correct source file, especially during forward label references from a PUT
file since another file may be the current text file when such errors are
detected.

Added pMainFile pointer to the Assembler object so that the original pTextFile
can be used to track the current text file being processed (which can be the
main source file or one included by a PUT directive.)  The pMainFile will
always point to the main source file opened at the time the Assembler object
was created.  When the end of pTextFile is encountered, the main assembler
routine will switch it back to pMainText if that wasn't already the current
source file being processed or flag the end of the assembly if it was the same.

Also added a pIncludedFiles linked list of TextFile nodes so that any other
TextFile that was allocated for any PUT directive can be maintained in memory
until the Assembler object is freed.