Implement USR opcode as used by Prince of Persia.
USR should be customizable by the developer but for now snap just hardcodes
this operator to work as it was used in the Prince of Persia source code.  It
was used in PoP to use the RW18 routines to write the object file output
directly to the disk at the required location.  In snap, it will be just saved
out to disk as a file like the SAV directive but with a different header that
can later be used by crackle to determine what side/track/offset/etc should be
used when placing the file in the final disk image.

Other changes with this commit include:
HEX directive had a bug where a comment at the end of a line with such a
directive would be throw off argument handling code and it would think that
too many HEX values were specified in the operand string.  This was fixed by
using the SizedString_EnumCurr() method to determine if all of the operand
string was consumed at the end of argument parsing.  I added a test which
exposed this bug in the HEX directive (by having a comment at the end of a HEX
test case.)  I also added similar tests for ASC, DB, and DA directives.

PUT directive should have thrown an error on nested PUT calls.