Implement DO, ELSE, FIN directives.
Now support conditional compilation.  Unlike the original Merlin, I didn't
bother to limit the conditional clauses to 8 levels and instead the nesting
level is only limited to memory available to allocate structs to be placed on
the conditional stack.

Changes include:

Now have LINEINFO_FLAG_FORWARD_REFERENCE flag in LineInfo structure so that
handlers such as the one for DO could know if they were being reprocessed for
an errant forward reference that occurred in its conditional expression.

Updated AssemblerBase test class's dupe() method to use a 512 character buffer
and assert that the string parameter length will fit in this 512 buffer.  Some
of the long strings used for the DO/ELSE/FIN testing caused an overflow in this
method which lead to corruption of the structures used for tracking memory
leaks.