Update sources from code review.
I code reviewed all of the snap and crackle code as I read through it
all to get back up to speed on where I left it.  During that code
review I cleaned up some of the code.  The biggest changes were taking
advantage of the new __try/__catch macroes I wrote awhile back which
use the setjmp and longjmp functionality to jump from a throw to the
neareset catch on the call stack.  With this new functionality, it is
no longer required that every caller between the throw location and the
catch handler with active code also have a __try/__catch block which
rethrows on exception.

I also took advantage of the allocateAndZero() function from utils.h
which allocates, throws on OOM, and zero fills from one convenient
function.