Missing FIN is now a warning.
I had to add warning support to snap.  Previously it would only output
error messages and only tracked the number of errors encountered.

When a missing FIN was encountered, the line number given as the cause
of the error was the last line of the source file.  This didn't really
help track down which DO was missing the FIN.  Now the line number for
the DO instruction is stored in the conditional structure so that when
this warning is generated, it can display the line number of the DO
which is actually missing the FIN.