Default parameter variables to 0.
Now default ]0 to ]9 variables to a zero value during initialization
of the assembler.

While testing this change, I hit an issue where I was corrupting the
g_pExceptionHandlers global by returning directly from within a __try
block instead of leaving it cleanly or throwing an exception.  I found
all the cases where I was doing this in the assembler and fixed them.
I also added an assert to the setup() method for all assembler tests to
catch such corruption earlier in the future.

I also had to update the out of memory failure test cases for the
Assembler init paths since more allocations are now performed in this
path since 10 symbols are added to the table.