Updates to SizedString module.
Added two new methods:
  SizedString_IsNull()
    Returns non-zero if pString is NULL or the string it points to is
    NULL.
  SizedString_strtoul()
    Similar to the strtoul() C library function except that it works
    with SizedStrings which don't have a NULL terminator but an
    explicit length instead.

I also fixed a bug in SizedString_strdup() that I recently encountered.
It turns out that the original implementation of this function wasn't
NULL terminating the returned string properly.  It was allocating
enough space for the NULL terminator but instead of setting it to NULL,
it just copied whatever character happened to be sitting after the last
character in the SizedString pString buffer.