TextFile_GetNextLine() now returns SizedString.
Previously this function returned a char* pointer.  I switched it to
returning a SizedString object so that the TextFile module would no
longer need to modify the input text to insert NULL terminators.  This
old modification behaviour made it hard to have another cursor on the
same text file object as I want to do for macro and looping support.

Switching TextFile_GetNextLine() to return a SizedString meant that all
other code which used the output from this call would also have to be
modified to work with SizedStrings.  This caused updates in the
following modules:
* LineInfo now uses lineText member instead of pLineText field to
    store the string for this line.
* ParseCSV now takes a SizedString as an input and returns SizedStrings
    as well for each field.  This means that ParseCSV had to be
    modified to not modify the input to insert NULL line terminators
    and instead return an array of SizedString pointers.
* ParseLine now takes a SizedString as input.  It has already be
    converted to output SizedStrings to make it easier for the listing
    to use the same input later.
* DiskImage had to be modified to use SizedStrings when reading in and
    parsing its script file.
* Assembler was modified to use SizedStrings for each line to assemble
    and to use the SizedStrings returned from ParseCSV module for the
    PUT include directories.  This module wasn't as impacted as the
    DiskImage since many parts of it had already been converted to use
    SizedStrings before to save memory and simplify the code.
* ListFile can now print out lines which are SizedStrings.