Allow space characters in expressions.
Previously the line parser would use whitespace as a delimiter to
terminate the operands field and separate it from any comment that
might be located on the line as well.  This meant that expressions
like the following would be assembled incorrectly since the last 2
characters will be truncated:
  lda #" "
I encountered this code generation bug when comparing Prince of Persia
object code assembled by snap to an image of the released disk.

The operand field will now only be truncated by the end of line or the
start of a ';' prefixed comment field.  This change required the
addressing mode, assembler, and expression modules to be updated to
cleanly handle and stop processing the operand if there were spaces at
the end.