Support forward reference to labels on DO 0 lines.
There was code like this in FRAMEADV.S where EditorDisk was set to 0:
 beq :stripe
 ...
:stripe do EditorDisk

The relative address in the assembled machine code for the BEQ
instruction was incorrect.  It turns out that the back patching was
getting ignored because the assembler incorrectly thought that the
BEQ line was inside the "DO 0" clause when it was being back patched.

I moved the conditional flags to the LineInfo structure so that each
line can track whether it was inside a DO clause when it is first
encountered so that back patching gets the correct state.