Add FileOpen() API to open existing files with case insensitivity.
Can be used from within snap instead of fopen() to open files on Linux
and not have the open fail if the existing filename uses different
case.  Without this snap can't assemble PoP sources on Linux sicne the
PUT directives in the source use lower case names but the files
are actually in upper case.