Update crackle's nibble file output.
Now uses excess bytes within a track to apply self-sync byte padding at
the beginning of a track.  Previously I had self-sync padding at the
end of the track for RW18 tracks and spread throughout the sectors for
RWTS16 tracks.

This change should make it possible for me to take the resulting nibble
images and write them out to disk.  .nib files actually contain more
data than what can possibly fit in the track of a real disk.  Having
this padding at the beginning of the track means that the excess
self-sync bytes will be what gets overwritten when the write of actual
data towards the end of a track loops around to the beginning of the
track.

I have created Prince of Persia 5.25" disk images with this updated
crackle utility and it still works under Virtual II emulator.